Interface Consolidation (BEWARE! major API change)
With this redefinition of the interfaces, the API hopefully gets less cluttered
and less surprising in most places.
We are now using "L" for the handle to the Lua script everywhere, instead of
conf in some places. Names are referred to as "key" now everywhere instead of
var in some places.
Global variables are obtained with aot_get_val, and the order of the arguments
is more consistent with the other retrieving routines. In a next step the
aot_get_val generic interface should cover also values in tables.