Fix uninitialized access in vector routines.
Doing things like

ErrCode = ibset(ErrCode, aoterr_NonExistent)

is dangerous if ErrCode is not already set (and, for ErrCode to be meaningful
given the error code definitions, set to 0 initially).  Change such lines to
(e.g.)

ErrCode = ibset(0, aoterr_NonExistent)

if that's the first time ErrCode can be set in the procedure or explicitly set
all ErrCode elements to 0 (depending on which is most convenient).